chore(meshupgrade): Refactor useStepper name
